기능 수정

+++ src/main/java/kakaopay/sprinkle/dto/ReceiveResponse.java
@@ -3,14 +3,16 @@ package kakaopay.sprinkle.dto;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Getter
 @NoArgsConstructor
 public class ReceiveResponse {
 
     private Long userId;
-    private int amount;
+    private BigDecimal amount;
 
-    public ReceiveResponse(Long userId, int amount) {
+    public ReceiveResponse(Long userId, BigDecimal amount) {
         this.userId = userId;
         this.amount = amount;
     }
