+++ src/main/java/kakaopay/sprinkle/dto/SprinkleRequest.java
@@ -3,14 +3,16 @@ package kakaopay.sprinkle.dto;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Getter
 @NoArgsConstructor
 public class SprinkleRequest {
 
-    private int amount;
+    private BigDecimal amount;
     private int numberOfRecipients;
 
-    public SprinkleRequest(int amount, int numberOfRecipients) {
+    public SprinkleRequest(BigDecimal amount, int numberOfRecipients) {
         this.amount = amount;
         this.numberOfRecipients = numberOfRecipients;
     }
